Fixed build on macOS
There was a circular dependency between
PortLinks.h and Port.cpp that was resolved
by removing PortLinks.h and including the
links functionality in Ports directly.
