+++ Dockerfile
@@ -1,6 +1,49 @@
 FROM ubuntu:18.04
-RUN apt-get update && apt-get install -y --no-install-recommends python=3.7 \
-     && apt-get clean \
-     && rm -rf /var/lib/apt/lists/*
-RUN wget https://repo.continuum.io/miniconda/Miniconda3-4.5.12-Linux-x86_64.sh && echo "e5e5b4cd2a918e0e96b395534222773f7241dc59d776db1b9f7fedfcb489157a  Miniconda3-4.5.12-Linux-x86_64.sh" | sha256sum -c --status && bash ./Miniconda3-4.5.12-Linux-x86_64.sh -document -p /root/miniconda && rm -f Miniconda3-4.5.12-Linux-x86_64.sh && echo "export PATH=\"/root/miniconda/bin:\$PATH\"" >> /root/.bashrc
+
+ENV LANG=C.UTF-8 LC_ALL=C.UTF-8
+ENV PATH /opt/conda/bin:$PATH
+
+RUN apt-get update --fix-missing && apt-get install -y wget bzip2 ca-certificates \
+    libglib2.0-0 libxext6 libsm6 libxrender1 \
+    git mercurial subversion
+
+########################################################
+# Essential packages for remote debugging and login in
+########################################################
+
+RUN apt-get update && apt-get upgrade -y && apt-get install -y \
+    apt-utils gcc g++ openssh-server cmake build-essential gdb gdbserver rsync vim
+
+RUN mkdir /var/run/sshd
+RUN echo 'root:root' | chpasswd
+RUN sed -i 's/PermitRootLogin prohibit-password/PermitRootLogin yes/' /etc/ssh/sshd_config
+
+# SSH login fix. Otherwise user is kicked off after login
+RUN sed 's@session\s*required\s*pam_loginuid.so@session optional pam_loginuid.so@g' -i /etc/pam.d/sshd
+
+ENV NOTVISIBLE "in users profile"
+RUN echo "export VISIBLE=now" >> /etc/profile
+
+# 22 for ssh server. 7777 for gdb server.
+EXPOSE 22 7777
+
+RUN useradd -ms /bin/bash debugger
+RUN echo 'debugger:pwd' | chpasswd
+
+########################################################
+# Add custom packages and development environment here
+########################################################
+
+RUN su debugger
+RUN cd /home/debugger
+RUN wget --quiet https://repo.anaconda.com/miniconda/Miniconda2-4.5.11-Linux-x86_64.sh -O ~/miniconda.sh && \
+    /bin/bash ~/miniconda.sh -b -p /opt/conda && \
+    rm ~/miniconda.sh && \
+    ln -s /opt/conda/etc/profile.d/conda.sh /etc/profile.d/conda.sh && \
+    echo ". /opt/conda/etc/profile.d/conda.sh" >> ~/.bashrc && \
+    echo "conda activate base" >> ~/.bashrc
 RUN conda update -y --all && conda install -y conda-build curl git expat openssl && conda config --add channels tpeulen && conda config --add channels conda-forge
+
+########################################################
+
+CMD ["/usr/sbin/sshd", "-D"]
