+++ docker-compose.yaml
@@ -1,0 +1,20 @@
+# From: https://github.com/shuhaoliu/docker-clion-dev/blob/master/docker-compose.yml
+
+version: '3'
+
+services:
+  gdbserver:
+    build:
+      context: ./
+      dockerfile: ./Dockerfile
+    image: clion_dev
+    security_opt:
+      - seccomp:unconfined
+    container_name: debug
+    ports:
+      - "7776:22"
+      - "7777:7777"
+    volumes:
+      - .:/home/debugger/code
+    working_dir: /home/debugger/code
+    hostname: debug
