+++ test/test_mongo_object.py
@@ -75,6 +75,7 @@ class Tests(unittest.TestCase):
         mo.set_array_int("i", [3, 4])
         self.assertEqual(mo.get_array_int("i"), (3, 4))
 
+    @unittest.expectedFailure
     def test_read_json(self):
         json_file = "./inputs/session_template.json"
